+++ webpack.config.js
@@ -71,6 +71,7 @@ const base = {
         { from: 'src/examples/remote-pdf-allowed.html', to: 'remote-pdf-allowed.html', transform: (content) => injectPartials(content, '.') },
         { from: 'src/examples/remote-pdf-cors-fail.html', to: 'remote-pdf-cors-fail.html', transform: (content) => injectPartials(content, '.') },
         { from: 'src/examples/kitchen-sink.html', to: 'kitchen-sink.html', transform: (content) => injectPartials(content, '.') },
+        { from: 'src/examples/multilingual.html', to: 'multilingual.html', transform: (content) => injectPartials(content, '.') },
         { from: 'src/examples/pdf-a-go-go-showcase.pdf', to: 'pdf-a-go-go-showcase.pdf' },
         { from: 'src/examples/example.pdf', to: 'example.pdf' },
         { from: 'src/examples/example_large.pdf', to: 'example_large.pdf' },